python_api/requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
pillow
python-multipart
openai
python-dotenv
